deleteHook 修正
